requirements.txt
ibm-watsonx-ai==2.0.2
python-dotenv